Ability to build road

+++ src/game/BaseEntity.java
@@ -20,7 +20,7 @@ public abstract class BaseEntity {
 
 	public int health, maxHealth;
 	public float offensiveStr, defensiveStr, rangedStr;
-	public int action = 1, maxAction = 1;
+	public double action = 1, maxAction = 1;
 	public int range = 0;
 
 	public City sortie = null; //If this unit was raised by a sortie order, this will hold the city it "belongs" to
